Fix how Intake handles multiValued entries. Initial patch to IntegerField provided by Derek Gaasch.
git-svn-id: https://svn.apache.org/repos/asf/jakarta/turbine/fulcrum/trunk@222998 13f79535-47bb-0310-9956-ffa450edef68

+++ intake/src/java/org/apache/fulcrum/intake/model/DoubleField.java
@@ -61,10 +61,8 @@ import org.apache.fulcrum.intake.validator.DoubleValidator;
 import org.apache.fulcrum.intake.xmlmodel.XmlField;
 
 /**
- * Creates Double Field objects.
+ * Processor for double fields.
  *
- * @author <a href="mailto:quintonm@bellsouth.net">Quinton McCombs</a>
- * @author <a href="mailto:seade@backstagetech.com.au">Scott Eade</a>
  * @version $Id$
  */
 public class DoubleField
@@ -153,18 +151,21 @@ public class DoubleField
         if (isMultiValued)
         {
             String[] inputs = parser.getStrings(getKey());
-            Double[] values = new Double[inputs.length];
+            double[] values = new double[inputs.length];
             for (int i = 0; i < inputs.length; i++)
             {
                 values[i] = StringUtils.isNotEmpty(inputs[i])
-                        ? new Double(inputs[i]) : (Double) getEmptyValue();
+                        ? new Double(inputs[i]).doubleValue() 
+                        : ((Double) getEmptyValue()).doubleValue();
             }
             setTestValue(values);
         }
         else
         {
             String val = parser.getString(getKey());
-            setTestValue(StringUtils.isNotEmpty(val) ? new Double(val) : (Double) getEmptyValue());
+            setTestValue(StringUtils.isNotEmpty(val) 
+                    ? new Double(val) : (Double) getEmptyValue());
         }
     }
+    
 }
