+++ intake/src/java/org/apache/fulcrum/intake/model/IntegerField.java
@@ -61,16 +61,13 @@ import org.apache.fulcrum.intake.validator.IntegerValidator;
 import org.apache.fulcrum.intake.xmlmodel.XmlField;
 
 /**
- * @author <a href="mailto:jmcnally@collab.net">John McNally</a>
- * @author <a href="mailto:hps@intermeta.de">Henning P. Schmiedehausen</a>
- * @author <a href="mailto:quintonm@bellsouth.net">Quinton McCombs</a>
- * @author <a href="mailto:seade@backstagetech.com.au">Scott Eade</a>
+ * Processor for int fields.
+ *
  * @version $Id$
  */
 public class IntegerField
         extends Field
 {
-
     /**
      * Constructor.
      *
@@ -154,18 +151,21 @@ public class IntegerField
         if (isMultiValued)
         {
             String[] inputs = parser.getStrings(getKey());
-            Integer[] values = new Integer[inputs.length];
+            int[] values = new int[inputs.length];
             for (int i = 0; i < inputs.length; i++)
             {
                 values[i] = StringUtils.isNotEmpty(inputs[i])
-                        ? new Integer(inputs[i]) : (Integer) getEmptyValue();
+                        ? new Integer(inputs[i]).intValue() 
+                        : ((Integer) getEmptyValue()).intValue();
             }
             setTestValue(values);
         }
         else
         {
             String val = parser.getString(getKey());
-            setTestValue(StringUtils.isNotEmpty(val) ? new Integer(val) : (Integer) getEmptyValue());
+            setTestValue(StringUtils.isNotEmpty(val) 
+                    ? new Integer(val) : (Integer) getEmptyValue());
         }
     }
+    
 }
